+++ nucs/examples/bacp/__main__.py
@@ -12,11 +12,11 @@
 ###############################################################################
 import json
 
-from nucs.constants import OPTIM_PRUNE
 from nucs.examples.bacp.bacp_problem import BACPProblem
 from nucs.examples.default_argument_parser import DefaultArgumentParser, run_optimizer, solver_kwargs_from_args
 from nucs.heuristics.heuristics import VAR_HEURISTIC_SMALLEST_DOMAIN
 from nucs.solvers.backtrack_solver import BacktrackSolver
+from nucs.solvers.solver import OPTIM_PRUNE
 
 # Run with the following command (the second run is much faster because the code has been compiled):
 # NUMBA_CACHE_DIR=.numba/cache python -m nucs.examples.bacp