+++ nucs/solvers/bc_algorithm.py
@@ -17,15 +17,9 @@ from numpy.typing import NDArray
 
 from nucs.buckets import STORAGE_OFFSET, buckets_add, buckets_pop
 from nucs.constants import (
-    CHOICE_POINT_TRAIL_MARK,
     DOMAIN_MAX,
     DOMAIN_MIN,
     EVENT_NB,
-    OFFSETS_PARAM,
-    OFFSETS_VARIABLE,
-    PROBLEM_BOUND,
-    PROBLEM_INCONSISTENT,
-    PROBLEM_UNBOUND,
     PROP_ENTAILMENT,
     PROP_INCONSISTENCY,
     STATS_ALG_IDX_FILTER_NB,
@@ -39,6 +33,14 @@ from nucs.constants import (
     STATS_MAX,
 )
 from nucs.numba_helper import ComputeDomainsFunctions
+from nucs.problems.problem import (
+    OFFSETS_PARAM,
+    OFFSETS_VARIABLE,
+    PROBLEM_BOUND,
+    PROBLEM_INCONSISTENT,
+    PROBLEM_UNBOUND,
+)
+from nucs.solvers.choice_points import CHOICE_POINT_TRAIL_MARK
 from nucs.solvers.state import tighten_at, trail_push, unbound_index